archive/postgresql-5container/dashboard/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
folium>=0.14.0
streamlit-folium>=0.15.0
plotly>=5.17.0
Pillow>=10.0.0
bcrypt>=4.0.0
opencv-python-headless>=4.8.0
Django>=4.2.0
psycopg2-binary>=2.9.0
google-generativeai>=0.3.0
ollama>=0.1.0
requests>=2.31.0

